fix: allow blur to take full screen on modal open
